zbd/006: reset only the test target zone
The test case zbd/006 performs random writes to the first sequential
write required zone of the test target zoned block device. To prepare
for this operation, it invokes the blkzone reset command, specifying the
offset of the test target zone. However, the zone count option is not
specified to the command. This resulted in reset of all sequential write
required zones on the device. This zone reset operation is unnecessary
for zones other than the first one and significantly increases the
operation time.

To address this issue, add the zone count option to the blkzone reset
command. Additionally, use long option names for better readability and
clarity.

Reviewed-by: Damien Le Moal <dlemoal@kernel.org>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>