nvme/{034,035}: skip when the test target namespace has metadata
According to the comment in nvmet_passthru_override_id_ns() in the
kernel code drivers/nvme/target/passthru.c, "Presently the NVMEof target
code does not support sending metadata, so we must disable it here".
Because of this limitation, when TEST_DEV namespace has the format with
metadata, the passthru target namespaces created in the test cases
nvme/034 and nvme/035 have zero capacity. This makes I/Os fail and the
test cases fail.

To avoid the failures, skip the test cases when TEST_DEV has metadata.
For that purpose, introduce the function _test_dev_has_no_metadata(),
which checks the sysfs attribute metadata_bytes.

Cc: Logan Gunthorpe <logang@deltatee.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>