common/null_blk: wait device ready in _configure_null_blk
Depending on test conditions, the test case block/017 fails when it is
run two or more times. It is quirky with load/unload null_blk in high
frequency. To avoid the failure, wait for device ready before testing
it.

block/017 (do I/O and check the inflight counter)            [passed]
    runtime  1.675s  ...  1.319s
block/017 (do I/O and check the inflight counter)            [failed]
    runtime  1.319s  ...  1.690s
    --- tests/block/017.out	2025-07-09 03:18:56.000000000 +0000
    +++ /usr/local/blktests/results/nodev/block/017.out.bad	2025-07-10 14:57:21.909726917 +0000
    @@ -4,9 +4,9 @@
     sysfs stat 1
     diskstats 1
     sysfs inflight reads 1
    -sysfs inflight writes 1
    -sysfs stat 2
    -diskstats 2
    +sysfs inflight writes 0
    ...
    (Run 'diff -u tests/block/017.out /usr/local/blktests/results/nodev/block/017.out.bad' to see the entire diff)

Signed-off-by: Frank Liang <xiliang@redhat.com>
[Shin'ichiro: improved the commit message]
Link: https://github.com/linux-blktests/blktests/pull/189
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>