scsi/011: allow "set -e" in run_test
The test case has the function run_test() which enables bash errexit
feature in it, so that command failures can be caught easily. However,
this errexit feature is not working because the call to run_test() is
followed by '||' operator. (ref: "set -e" description in man bash)

To enable the bash errexit feature in the call to run_test(), do not use
the '||' operator. Instead, check the exit value of the run_test() using
the special parameter '$?'. This triggers the shellcheck warning SC2181
then suppress it.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>