md/002: use --run option instead of "echo y"
When mdadm creates a new array, it asks questions to users in the
console, which is not suitable for automated test script runs. Then the
test case md/002 used "echo y" with pipe to answer the questions
automatically.

However, starting with mdadm version 4.4, the "echo y" approach no
longer works effectively, especially after a commit which introduced a new
question [1]. This caused the test case failure.

[1] https://git.kernel.org/pub/scm/utils/mdadm/mdadm.git/commit/?h=main&id=e97c4e18c847803016aa60066cb6e57c528d83a6

The --run option is now the recommended alternative because it directly
bypasses the questions. To avoid the failure, add the "--run" option to
the mdadm create commands instead of the "echo y".

Link: https://github.com/linux-blktests/blktests/issues/201
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>