check, new: introduce test_device_array()
As to the test target devices defined in TEST_DEVS variable, blktests
assumes that each test case with test_device() function is run for each
single device defined in TEST_DEVS. On the other hand, it is suggested
to support a test case for not a single device but multiple devices.

To fulfill this request, introduce a new test function
test_device_array() and TEST_CASE_DEV_ARRAY. TEST_CASE_DEV_ARRAY is an
associative array. Test case names are the keys, and the lists of block
devices are the values of the associative array. When a test case
defines test_device_array() and users specify TEST_CASE_DEV_ARRAY in
the config file for the test case, the test_device_array() is called.
Blktests framework checks that each device in TEST_CASE_DEV_ARRAY is a
block device, then call device_requires() and group_device_requires()
for it. Blktests prepares TEST_DEV_ARRAY and TEST_DEV_ARRAY_SYSFS_DIRS
which contain the list of devices and corresponding sysfs paths, so that
test_device_array() can refer to these to run the test.

Tested-by: John Garry <john.g.garry@oracle.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>