common/rc, dm/003, nvme/065: move TO_SKIP to common/rc
TO_SKIP is defined as readonly at two places in dm/003 and nvme/065.
When these two test cases are run at once, bash complains that the
readonly value is assigned twice. To avoid the error, do not declare the
value as readonly. Also, move the declarations of TO_SKIP to common/rc
to avoid the duplicated declarations.

Reported-by: Yi Zhang <yi.zhang@redhat.com>
Link: https://github.com/linux-blktests/blktests/pull/204
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>