Makefile: follow symbolic links of *.out files
Currently, "make check" uses a find command to ensure that '*.out' files
do not have execute permission. However, when an .out file is a symbolic
link, it always has execute permission regardless of the true permission
of the linked file and the "make check" command fails. This is false-
alert.

To avoid the false-alert, add -L option to the find command so that it
follows symbolic links. With this, the execute permission check is done
not for the symbolic links but for the linked files.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>