throtl/rc,002: adjust to scsi_debug
Currently, the test case throtl/002 assumes that the test target device
is null_blk. For null_blk, it is possible to specify the page size as
the maximum I/O size during set up. However, in the case of scsi_debug,
the maximum I/O size is not configurable, causing the test case to fail
when the test case is extended to support scsi_debug.

To prevent the failure, set the maximum I/O size after the test device
is initialized, regardless of the device type. To achieve this, add a
new helper function _throtl_set_max_io_size() which modifies the sysfs
attribute max_sectors_kb to adjust the maximum I/O size for the device.

Suggested-by: Yu Kuai <yukuai1@huaweicloud.com>
Reviewed-by: Chaitanya Kulkarni <kch@nvidia.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>