throtl/004: adjust to scsi_debug
Currently, the test case throtl/004 assumes that the test target device
is null_blk. When it verifies the dd command error message, it checks
the string "/dev/dev_nullb" also. However, this test will fail when the
test case is extended to support scsi_debug.

To avoid the failure, remove dependency on the specific device name.
Save the dd command error output to the FULL file, and check only the
error message part in the FULL file.

Reviewed-by: Chaitanya Kulkarni <kch@nvidia.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>