blktrace: add blktrace zone management regression test
Create a new blktrace test group and add a regression test for a
blktrace false positive WARNING that occurs when zone management
commands are traced with blktrace on V1 version.

Bug: https://syzkaller.appspot.com/bug?extid=153e64c0aa875d7e4c37
Location: kernel/trace/blktrace.c:367-368

The test:
1. Creates a zoned null_blk device (8 zones, 1GB, no conventional zones)
2. Starts blktrace on the device
3. Issues zone open command for all zones
4. Checks dmesg for the false positive WARNING

Device configuration:
- Total size: 1GB
- Zone size: 128MB
- Number of zones: 8
- Conventional zones: 0

If the WARNING is found, the bug is present and logged to the full
output. If no WARNING appears, the bug is fixed.

Note: The bug uses WARN_ON_ONCE, so it triggers only once per boot.
Subsequent runs after the first trigger will not show the WARNING:
commit 4a0940bdcac260be1e3460e99464fa63d317c6a2
Author: Chaitanya Kulkarni <ckulkarnilinux@gmail.com>
Date:   Mon Oct 27 19:46:19 2025 -0700

    blktrace: use debug print to report dropped events

https://lore.kernel.org/linux-block/20251028024619.2906-1-ckulkarnilinux@gmail.com/

Signed-off-by: Chaitanya Kulkarni <ckulkarnilinux@gmail.com>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>