md/{002,004}: adjust per_host_store to use numeric boolean
In both md/004 and md/002, when loading the scsi_debug module,
"true" is passed as the value for per_host_store. However, on
older kernels, kstrtobool is more limited, so we get the error:
"scsi_debug: `true' invalid for parameter `per_host_store'".
Change the boolean from "true" to "1" to avoid this issue.

Signed-off-by: John Pittman <jpittman@redhat.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>